+++ packages/riverpod/test/providers/future_provider/auto_dispose_future_provider_test.dart
@@ -25,14 +25,14 @@ void main() {
 
     expect(
       ref.state,
-      const AsyncLoading<int>(),
+      const AsyncData<int>(0, isRefreshing: true),
     );
 
     verifyOnly(
       listener,
       listener(
         const AsyncData(0),
-        const AsyncLoading<int>(),
+        const AsyncData<int>(0, isRefreshing: true),
       ),
     );
   });
@@ -107,7 +107,7 @@ void main() {
 
     verifyOnly(
       listener,
-      listener(null, const AsyncLoading<int>()),
+      listener(null, const AsyncData<int>(42, isRefreshing: true)),
     );
 
     container.read(dep.state).state = Future.value(21);
@@ -124,6 +124,36 @@ void main() {
     );
   });
 
+  test('can refresh .future', () async {
+    var future = Future.value(1);
+    final provider = FutureProvider.autoDispose((ref) => future);
+    final container = createContainer();
+
+    container.listen(provider.future, (prev, value) {});
+
+    expect(await container.read(provider.future), 1);
+
+    future = Future.value(42);
+
+    expect(await container.refresh(provider.future), 42);
+    expect(container.read(provider), const AsyncData(42));
+  });
+
+  test('can refresh .stream', () async {
+    var future = Future.value(1);
+    final provider = FutureProvider.autoDispose((ref) => future);
+    final container = createContainer();
+
+    container.listen(provider.stream, (prev, value) {});
+
+    expect(await container.read(provider.stream).first, 1);
+
+    future = Future.value(42);
+
+    expect(await container.refresh(provider.stream).first, 42);
+    expect(container.read(provider), const AsyncData(42));
+  });
+
   test('can be refreshed', () async {
     var result = 0;
     final container = createContainer();
@@ -137,7 +167,7 @@ void main() {
     result = 1;
     expect(
       container.refresh(provider),
-      const AsyncValue<int>.loading(),
+      const AsyncValue<int>.data(0, isRefreshing: true),
     );
 
     expect(await container.read(provider.future), 1);