+++ website/static/snippets/declare.dart
@@ -1,3 +1,10 @@
+// ignore_for_file: use_key_in_widget_constructors
+
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+/* SNIPPET START */
+
 // A shared state that can be accessed by multiple
 // objects at the same time
 final countProvider = StateProvider((ref) => 0);
@@ -6,7 +13,7 @@ final countProvider = StateProvider((ref) => 0);
 class Title extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final count = ref.watch(countProvider).state;
+    final count = ref.watch(countProvider);
     return Text('$count');
   }
 }